% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_data.R
\name{arc_data}
\alias{arc_data}
\title{Copy files from data directory to archive directory}
\usage{
arc_data(asmbdat_path = getwd(), archive_path = NULL, ...)
}
\arguments{
\item{asmbdat_path}{path to projects asmbdat directory. Defaults to working
directory}

\item{archive_path}{path to projects archive directory. Will create a new
archive directory if one doesn't exists as well as dated folders inside
archive directory. Dated archive directory defaults to System Date}

\item{...}{arguments passed to fs::dir_info() i.e. regexp, glob}
}
\value{
files moved to data archive directory.
}
\description{
Copy files from data directory to archive directory
}
