% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{as_gitlab_api_url}
\alias{as_gitlab_api_url}
\title{Convert a GitLab URL to its API format}
\usage{
as_gitlab_api_url(url)
}
\arguments{
\item{url}{URL of a GitLab resource}
}
\value{
\code{character} URL with attributes for host, group, project,
  branch, file, and whether it is a tree (is_tree).
}
\description{
The recursive option is set to \code{true} for trees
}
\keyword{internal}
