% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_standards.R
\name{check_header}
\alias{check_header}
\title{Check if an R script has an appropriate Cognigen header}
\usage{
check_header(x, path)
}
\arguments{
\item{x}{a scanned R script}

\item{path}{a \code{character} containing a directory path}
}
\value{
a named \code{list} with "pass" and "reason" elements
}
\description{
This checks that the script has a Cognigen header, defines a name that
matches the file path, has a copyright statement, and defines the script's
purpose.
}
\keyword{internal}
