% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_standards.R
\name{check_rout}
\alias{check_rout}
\title{Determines if an R script should have an associated Rout file, then checks
the content of the Rout}
\usage{
check_rout(x, path)
}
\arguments{
\item{x}{a scanned R script}

\item{path}{a \code{character} containing a directory path}
}
\value{
a named \code{list} with "pass" and "reason" elements
}
\description{
For any R file, there should either be an associated Rout file or
documentation in the header stating that the script is sourced elsewhere.
This checks that one and only one of these conditions is met. If there is an
associated Rout file, this also checks that the R version in the header
matches the R version in the Rout.
}
\keyword{internal}
