% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_standards.R
\name{check_session_info}
\alias{check_session_info}
\title{Check if a scanned R script calls \code{\link[utils]{sessionInfo}} or
\code{\link[devtools]{session_info}} as the final call of the script}
\usage{
check_session_info(x, path)
}
\arguments{
\item{x}{a scanned R script}

\item{path}{a \code{character} containing a directory path}
}
\value{
a named \code{list} with "pass" and "reason" elements
}
\description{
Check if a scanned R script calls \code{\link[utils]{sessionInfo}} or
\code{\link[devtools]{session_info}} as the final call of the script
}
\keyword{internal}
