% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_path.R
\name{clean_path}
\alias{clean_path}
\title{Clean file paths}
\usage{
clean_path(path)
}
\arguments{
\item{path}{file paths. Defaults to the path of the source editor context.}
}
\value{
\code{character} vector of cleaned file paths.
}
\description{
Removes quote characters, replaces backslashes with forward slashes, re-lists
paths if multiple paths are stored in one element of the vector.
}
\keyword{internal}
