% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Redit.R
\name{clean_version}
\alias{clean_version}
\title{Clean a supported R version to its \code{R.version.string} representation}
\usage{
clean_version(version)
}
\arguments{
\item{version}{either \code{NULL} for the current R version, or a
\code{character} in the form \code{"N.n.n"} or \code{"Nnn"}. Ignored with a
warning for files that already have a header.}
}
\value{
\code{character} in the form of \code{R.version.string}
}
\description{
Clean a supported R version to its \code{R.version.string} representation
}
\keyword{internal}
