% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_path.R
\name{convert_path}
\alias{convert_path}
\title{Facilitator for \code{path_to_unix} and \code{path_to_windows}}
\usage{
convert_path(path = NULL, ask = FALSE, normalize = TRUE, to = "unix")
}
\arguments{
\item{path}{file paths. Defaults to the path of the source editor context.}

\item{ask}{\code{logical} indicating whether to interactively ask for input.
This is particularly helpful for file paths including single backslashes
(Windows paths).}

\item{normalize}{\code{logical} indicating whether to normalize the file
paths. Ignored on platforms other than the desired OS. When \code{TRUE},
files that do not exist result in \code{NA}.}

\item{to}{which OS representation to match. Either 'unix' or 'windows'.}
}
\value{
\code{character} vector with length of \code{path} containing the
  converted paths.
}
\description{
Facilitator for \code{path_to_unix} and \code{path_to_windows}
}
\keyword{internal}
