% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_copy.R
\name{cp}
\alias{cp}
\title{Execute \code{cp} command}
\usage{
cp(from, to, args = NULL)
}
\arguments{
\item{from}{a file name, path, or directory.}

\item{to}{a file name, path, or directory.}

\item{args}{\code{character} vector of additional arguments to \code{cp}. The
default includes \code{"-p"}, to preserve mode, ownership, and timestamps.
See \code{file_copy(args = "help")} for \code{cp} arguments.}
}
\value{
\code{logical} indicating if copy operation succeeded.
}
\description{
Execute \code{cp} command
}
\keyword{internal}
