% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{curl_and_parse}
\alias{curl_and_parse}
\title{Run the \code{curl} shell command and parse the JSON result}
\usage{
curl_and_parse(url)
}
\arguments{
\item{url}{URL of a GitLab resource.}
}
\value{
\code{list} of information about the resource.
}
\description{
Run the \code{curl} shell command and parse the JSON result
}
\keyword{internal}
