% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcb.R
\name{.rcb}
\alias{.rcb}
\title{Execute an R script in batch mode}
\usage{
.rcb(path)
}
\arguments{
\item{path}{file path of R program.}
}
\value{
\code{logical} indicating whether the program executed successfully.
}
\description{
Execute an R script in batch mode
}
\keyword{internal}
