% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{download_gitlab}
\alias{download_gitlab}
\title{Download trees (directories) and blobs (files) from GitLab using the GitLab
API}
\usage{
download_gitlab(url, destfile = NULL)
}
\arguments{
\item{url}{a single web URL pointing to a GitLab resource}

\item{destfile}{either a single non-existing directory or a single file.
Defaults to the basename of the \code{from} file.}
}
\value{
\code{logical} indicating whether resources were successfully
  downloaded. This function does report success even if some resources are
  not successfully downloaded, with a warning for each resource that could
  not be downloaded.
}
\description{
\code{url} is converted to its GitLab API format. For tree
  input, the first 100 pages are downloaded. For blob input, the blob is
  downloaded. Version 4 of the GitLab API is used.
}
