% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{download_gitlab_tree}
\alias{download_gitlab_tree}
\title{Helper to \code{download_gitlab} to facilitate traversing through trees}
\usage{
download_gitlab_tree(api_url, destfile, curl_result)
}
\arguments{
\item{api_url}{a single GitLab API URL pointing to a tree}

\item{destfile}{a single non-existing directory}

\item{curl_result}{\code{list} result of the parsed \code{curl} call to
\code{api_url}}
}
\value{
created needed directories, calls \code{download_gitlab} for each
  blob / file path pair, then invisibly returns \code{NULL}
}
\description{
Helper to \code{download_gitlab} to facilitate traversing through trees
}
\keyword{internal}
