% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_copy.R
\name{file_copy}
\alias{file_copy}
\title{Copy files and directories using Unix \code{cp} command}
\usage{
file_copy(from, to, args = c("-p", "-r"))
}
\arguments{
\item{from}{\code{character} vector containing file names or paths. Can also
contain directories, but \code{args} must include \code{"-r"} (recursive).}

\item{to}{\code{character} vector containing file names, paths, or existing
directories.}

\item{args}{\code{character} vector of additional arguments to \code{cp}. The
default includes \code{"-p"}, to preserve mode, ownership, and timestamps;
and \code{"-r"} to copy directories recursively. See \code{file_copy(args =
"help")} for \code{cp} arguments.}
}
\value{
returns a \code{logical} vector indicating which copy operations
  succeeded for each of the files attempted.
}
\description{
This function is designed to use the Unix \code{cp} command, but will revert
to \code{base::file.copy()} in Windows environments. In this case, including
\code{"-r"} in \code{args} results in \code{recursive = TRUE}.
}
\examples{
\dontrun{
# Get cp help to review optional args
file_copy(args = "help")

# Copy a single file
file_copy("old-dir/file.R", "new-dir/file.R")

# Copy multiple files to an existing directory
file_copy(c("old-dir/file1.R", "old-dir/file2.R"), "new-dir/")

# Pass no optional arguments
file_copy("old-dir/file.R", "new-dir/file.R", args = NULL)

# Pass optional arguments
file_copy("old-dir/file.R", "new-dir/file.R", args = c("-b", "-n", "-p"))
}
}
\keyword{internal}
