% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_open.R
\name{file_open}
\alias{file_open}
\title{Open files according to their extensions or show a directory}
\usage{
file_open(..., no_copies = FALSE)
}
\arguments{
\item{...}{file paths to be opened.}

\item{no_copies}{\code{logical} indicating whether to avoid downloading
copies of files. If \code{TRUE} and some file cannot be opened by this
system, the Windows path is printed as a message.}
}
\value{
invisibly returns \code{NULL}.
}
\description{
The files will be opened if possible. Up to one directory can be shown in the
RStudio Files pane. 

The behavior for files that cannot be displayed in the RStudio source editor
or a browser is determined by \code{no_copies}. When \code{FALSE} (the
default), a read-only copy will be opened. When \code{TRUE}, the Windows file
path is printed as a message.
}
