% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_refresh.R
\name{file_refresh}
\alias{file_refresh}
\title{Refresh the active file in the RStudio source editor without saving}
\usage{
file_refresh()
}
\value{
invisibly returns \code{NULL}
}
\description{
Files updated outside of RStudio do not refresh immediately, even with
\code{rstudioapi::navigateToFile()}. This closes then opens the active file
in the RStudio source editor.
}
\keyword{internal}
