% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_snapshots.R
\name{file_snapshots}
\alias{file_snapshots}
\alias{diff_snapshot}
\alias{restore_snapshot}
\title{Interface to Cognigen zfs snapshots}
\usage{
file_snapshots(path = NULL, date = Sys.Date(), time = NULL)

diff_snapshot(path = NULL, date = Sys.Date(), time = "last")

restore_snapshot(
  path = NULL,
  date = Sys.Date(),
  time = "last",
  args = c("-p", "-r"),
  new_path = NULL,
  diff = TRUE
)
}
\arguments{
\item{path}{a file name or path. Defaults to the path of the source editor
context.}

\item{date}{a \code{character} in YYYY-MM-DD format or \code{Date} specifying
the date range to subset to. Complete dates are not required;
\code{character} representations of years or months are accepted. Defaults
to the current date.}

\item{time}{\code{character} assumed to be in 24-hour format; will be
adjusted if "AM" or "PM" is included. If a time is provided, the closest
matching snapshot time is selected. "first" and "last" are also accepted.
When more than a single \code{date} is provided, \code{time} should be
either NULL, "first", or "last".}

\item{args}{\code{character} vector of additional arguments to \code{cp}. The
default includes \code{"-p"}, to preserve mode, ownership, and timestamps;
and \code{"-r"} to copy directories recursively. See \code{file_copy(args =
"help")} for \code{cp} arguments.}

\item{new_path}{an optional file name or path indicating where to save the
restored snapshot. If provided, the original \code{path} will not be
modified.}

\item{diff}{\code{logical} indicating whether to call
\code{\link{diff_snapshot}}}
}
\value{
\code{file_snapshots} -- a \code{character} vector of file snapshots for a
given \code{date}. If \code{time} is provided, no more than one file snapshot
will be returned.

\code{diff_snapshot} -- invisibly returns \code{NULL}

\code{restore_snapshot} -- a \code{logical} indicating whether the copy
command was successful
}
\description{
See
\href{https://wiki.cognigencorp.com/display/Office/DIY+File+Restoration}{Wiki
DIY File Restoration}

\code{file_snapshots} gets the available paths to snapshots of a file on a
given snapshot date, month, or year.

\code{diff_snapshot} produces a diff of a file and a single snapshot. If
RStudio is available, the diff is rendered in the Viewer. Otherwise, a Unix
\code{diff} is printed to the console.

\code{restore_snapshot} restores a file using a single snapshot. By default,
this function restores a file to its most recent snapshot, saves the current
version in \code{tempdir()}, and outputs the result of \code{diff_snapshot}.
Directories cannot be restored.
}
\examples{
\dontrun{
file_snapshots(path = "script.R", date = Sys.date())
file_snapshots(date = "2020-01-01")

# check the differences between the latest snapshot and the current
# version of the current source document
diff_snapshot()

# restore the current source document to the most recent snapshot
restore_snapshot()

# restore a file to a snapshot on a given date, closest to a given time
restore_snapshot("file.R", date = "2020-02-19", time = "10:30")

# copy a restored file to a different directory or file
restore_snapshot("file.R", new_path = "archive/")
}
}
