% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_path.R
\name{fs_path}
\alias{fs_path}
\alias{path_sponsor}
\alias{path_drug}
\alias{path_project}
\alias{browse_project_email}
\alias{browse_project_sharepoint}
\title{Build file paths based on Cognigen's file system and directory structure.}
\usage{
path_sponsor(sponsor, path = ".")

path_drug(drug, sponsor = "*", path = ".")

path_project(project_number, drug = "*", sponsor = "*", path = ".")

browse_project_email(path = ".")

browse_project_sharepoint(path = ".")
}
\arguments{
\item{sponsor, drug, project_number}{\code{character} vectors of directory
patterns or names. Wildcard expansion (also known as 'globbing') is
supported.}

\item{path}{\code{character} vector of file or directory paths that will be
truncated to the project number, drug, or sponsor level. Ignored when the
first argument is provided.}
}
\value{
The \code{path_*} functions return the path to the sponsor, drug, or project
directory corresponding to either: the \code{path} of a file or directory; or
the name of the sponsor, drug, or project number.

The \code{browse_project_*} functions open the corresponding project's email
alias or SharePoint site in a new browser tab.

\code{path_sponsor()} returns the path to the sponsor directory.

\code{path_drug()} returns the path to the drug directory.

\code{path_project()} returns the path to the project directory.
}
\description{
These functions are only intended to be used at Cognigen.
}
\examples{
\dontrun{
# the below examples reference a test directory on Cognigen's file system.
cd("/misc/dceuticals/doloxan/009002/")

# the default behavior is to return the path to the sponsor, drug, or project.
path_sponsor()
path_drug()
path_project()

# or open the resource for the corresponding project.
# note that these will not produce desired results with the test dceuticals
# directory because the project alias and SharePoint site do not exist.
browse_project_email()
browse_project_sharepoint()

# the below examples do not rely on the working directory
cd("~")

# use the pattern "*" to match any sponsor, drug, or project number.

# get paths to all sponsor directories:
path_sponsor("*")

# get paths to sponsor directories that begin with "bio":
path_sponsor("bio*")

# get paths to drug directories that end with "umab":
path_drug("*umab")

# get paths to project directories that begin with "20" for drugs that end
# with "umab" or start with "ab":
path_project("20*", drug = c("*umab", "ab*"))

# open email aliases for all projects where the project number starts with
# "006" and ends with "58"
browse_project_email(path_project("006*58"))

}
}
