% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{get_gitlab_project_contents}
\alias{get_gitlab_project_contents}
\title{List of lists of all project contents}
\usage{
get_gitlab_project_contents(url)
}
\description{
List of lists of all project contents
}
\keyword{internal}
