% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{get_gitlab_project_id}
\alias{get_gitlab_project_id}
\title{Get a GitLab project id}
\usage{
get_gitlab_project_id(host, group, project)
}
\arguments{
\item{host}{GitLab API host to use.}

\item{group}{GitLab group name.}

\item{project}{GitLab project name.}
}
\value{
\code{integer} project id
}
\description{
Get a GitLab project id
}
\keyword{internal}
