% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{get_header_content}
\alias{get_header_content}
\alias{get_header_name}
\alias{get_header_history}
\alias{get_header_version}
\alias{get_header_copyright}
\alias{get_header_purpose}
\alias{get_header_input_files}
\alias{get_header_output_files}
\title{Get content from an R script header}
\usage{
get_header_name(path = NULL)

get_header_history(path = NULL)

get_header_version(path = NULL)

get_header_copyright(path = NULL)

get_header_purpose(path = NULL)

get_header_input_files(path = NULL)

get_header_output_files(path = NULL)
}
\arguments{
\item{path}{a file path to an R script. Defaults to the path of the source
editor context.}
}
\value{
\code{character} vector of requested header item.
}
\description{
For `get_header_input_files` and `get_header_output files`, file paths are
split by new lines, commas, and semicolons.
}
