% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_functions.R
\name{get_package_functions}
\alias{get_package_functions}
\alias{get_package_function_arguments}
\title{Get the names and/or arguments of the functions included in a package.}
\usage{
get_package_functions(package, unexported = FALSE, invisible = FALSE)

get_package_function_arguments(package, unexported = FALSE, invisible = FALSE)
}
\arguments{
\item{package}{a package name as a \code{character}}

\item{unexported}{\code{logical} indicating whether to include unexported
functions}

\item{invisible}{\code{logical} indicating whether to include unexported
functions that start with a \code{"."}. Ignored unless \code{unexported} is
\code{TRUE}}
}
\value{
\code{get_package_functions} -- returns a \code{character} vector of function
names.

\code{get_package_function_arguments} -- returns named \code{character} of
function arguments.
}
\description{
Get the names and/or arguments of the functions included in a package.
}
\examples{
\dontrun{
get_package_functions("utilscognigen")
get_package_function_arguments("utilscognigen")
}
}
\keyword{internal}
