% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_source_file.R
\name{get_source_file}
\alias{get_source_file}
\title{Get the path of the RStudio source editor or the R script being executed}
\usage{
get_source_file()
}
\value{
returns the path of the source editor or the script being executed.
}
\description{
For interactive sessions: Calls
\code{\link[rstudioapi]{getSourceEditorContext}} to obtain the source
context. The source editor context refers to the active file opened in
RStudio.

For non-interactive sessions: Parses \code{\link[base]{commandArgs}} to 
obtain the name of the script being executed.
}
