% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Redit.R
\name{get_user_full_name}
\alias{get_user_full_name}
\title{Get the full name of a \code{user}}
\usage{
get_user_full_name(user = Sys.getenv("USER"))
}
\arguments{
\item{user}{\code{character} user. Defaults to the system environment
variable \code{USER}.}
}
\value{
\code{character} full name of the \code{user}.
}
\description{
Get the full name of a \code{user}
}
\keyword{internal}
