% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_cognigen.R
\name{is_cognigen}
\alias{is_cognigen}
\title{Check if the current R session is at Cognigen.}
\usage{
is_cognigen()
}
\value{
logical
}
\description{
Check if the current R session is at Cognigen.
}
\keyword{internal}
