% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{is_gitlab_api_url}
\alias{is_gitlab_api_url}
\title{Check if a URL is a GitLab API URL}
\usage{
is_gitlab_api_url(url)
}
\arguments{
\item{url}{URL of a GitLab resource.}
}
\value{
\code{logical}
}
\description{
Check if a URL is a GitLab API URL
}
\keyword{internal}
