% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_cognigen.R
\name{is_slp_any}
\alias{is_slp_any}
\title{Check if the current R session is at any Simulations Plus division.}
\usage{
is_slp_any()
}
\value{
logical
}
\description{
Check if the current R session is at any Simulations Plus division.
}
\keyword{internal}
