% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{list_gitlab_projects}
\alias{list_gitlab_projects}
\title{List GitLab projects, optionally subset to groups}
\usage{
list_gitlab_projects(host = "gitlab.cognigencorp.com", groups = NULL)
}
\arguments{
\item{host}{GitLab API host to use.}

\item{groups}{GitLab group name(s).}
}
\value{
JSON \code{list} of GitLab projects.
}
\description{
List GitLab projects, optionally subset to groups
}
\keyword{internal}
