% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile.R
\name{logfile}
\alias{logfile}
\title{Open R, Rmd, SAS, tex, Rnw log files in RStudio for the current source
document}
\usage{
logfile(...)
}
\arguments{
\item{...}{file paths to open the log files of. Defaults to the path of the
source editor context.}
}
\value{
invisibly returns \code{NULL}
}
\description{
Open R, Rmd, SAS, tex, Rnw log files in RStudio for the current source
document
}
