% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Redit.R
\name{make_copyright_line}
\alias{make_copyright_line}
\title{Make the header lines for a copyright statement}
\usage{
make_copyright_line(copyright_holder = NULL)
}
\value{
\code{character} or \code{NULL}
}
\description{
Make the header lines for a copyright statement
}
\keyword{internal}
