% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Redit.R
\name{make_date_user_line}
\alias{make_date_user_line}
\title{Make the header line containing the date and user}
\usage{
make_date_user_line()
}
\value{
\code{character}
}
\description{
Make the header line containing the date and user
}
\keyword{internal}
