% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Redit.R
\name{make_header_files}
\alias{make_header_files}
\title{Make the header lines containing the input or output files}
\usage{
make_header_files(files, type = c("input", "output"))
}
\arguments{
\item{files}{input or output files.}

\item{type}{either \code{"input"} or \code{"output"}.}
}
\value{
\code{character} or \code{NULL}
}
\description{
Make the header lines containing the input or output files
}
\keyword{internal}
