% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_path.R
\name{make_windows_replacements}
\alias{make_windows_replacements}
\title{Make replacements to file paths to get the Windows representation}
\usage{
make_windows_replacements(path, normalize)
}
\arguments{
\item{path}{file paths. Defaults to the path of the source editor context.}

\item{normalize}{\code{logical} indicating whether to normalize the file
paths. Ignored on platforms other than the desired OS. When \code{TRUE},
files that do not exist result in \code{NA}.}
}
\value{
\code{character} vector the length of \code{path} containing the
  converted paths.
}
\description{
Make replacements to file paths to get the Windows representation
}
\keyword{internal}
