% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{parse_header}
\alias{parse_header}
\alias{parse_header_name}
\alias{parse_header_all_names}
\alias{parse_header_timestamp}
\alias{parse_header_version}
\alias{parse_header_copyright}
\alias{parse_header_purpose}
\alias{parse_header_input_files}
\alias{parse_header_output_files}
\alias{parse_header_is_sourced}
\title{Parse R script headers}
\usage{
parse_header(path = NULL, sections = .valid_header_sections, ...)

parse_header_name(header)

parse_header_all_names(header)

parse_header_timestamp(header)

parse_header_version(header)

parse_header_copyright(header)

parse_header_purpose(header)

parse_header_input_files(path, header)

parse_header_output_files(path, header)

parse_header_is_sourced(header)
}
\arguments{
\item{path}{a file path to an R script. Defaults to the path of the source
editor context.}

\item{sections}{header sections to parse}

\item{...}{additional arguments passed to \code{\link{get_header}}}

\item{header}{a header from an R script}
}
\value{
\code{parse_header} returns a \code{list} of parsed sections. Each
  \code{parse_<section>} function returns the parsed section.
}
\description{
Parse R script headers
}
\keyword{internal}
