% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcb.R
\name{rcb}
\alias{rcb}
\title{Execute R programs in batch mode}
\usage{
rcb(..., scanlogs = TRUE, as_job = getOption("utilscognigen.rcb_as_job", TRUE))
}
\arguments{
\item{...}{file paths of R programs. Defaults to the path of the source
editor context.}

\item{scanlogs}{\code{logical} indicating whether to execute
\code{\link{scanlogs}} on Rout log files.}

\item{as_job}{\code{logical} indicating whether to execute as an RStudio job.}
}
\value{
When executed as an RStudio job, \code{rcb} invisibly returns
  \code{NULL} after submitting the job. Otherwise returns a \code{logical}
  vector indicating whether each program executed successfully.
}
\description{
Calls \code{\link[callr]{rcmd}} on all files from each files'
parent directory. If the execution of any program is unsuccessful, later
programs are not executed. Command line equivalent for each run is \code{R
CMD BATCH --no-save --no-environ --no-init-file --no-restore script.R}.

In RStudio sessions, programs can be executed as jobs so the R Console
remains available.
}
\examples{
\dontrun{
rcb("script1.R", "script2.R")
}
}
