% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_io.R
\name{record_io}
\alias{record_io}
\title{Utility function supporting \code{record_input} and \code{record_output}}
\usage{
record_io(call, file, type = "input", quiet = FALSE)
}
\arguments{
\item{call}{a function call to read a data file or create a new file;
alternatively, call could be a simple file path.}

\item{file}{a file path.}

\item{type}{either 'input' or 'output' to document whether the recorded
information corresponds to an input or output file.}

\item{quiet}{a logical value indicating whether messages should be printed
(FALSE; default) or not (TRUE).}
}
\description{
Utility function supporting \code{record_input} and \code{record_output}
}
\keyword{internal}
