% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Redit.R
\name{Redit}
\alias{Redit}
\title{Create, update, and open R programs with QMS approved headers.}
\usage{
Redit(
  ...,
  version = NULL,
  purpose = NULL,
  input_files = NULL,
  output_files = NULL,
  backup = FALSE,
  open = rstudioapi::isAvailable()
)
}
\arguments{
\item{...}{file paths of R or Rmd files. Files without extensions are set to
\code{".R"}. Defaults to the path of the source editor context.}

\item{version}{either \code{NULL} for the current R version, or a
\code{character} in the form \code{"N.n.n"} or \code{"Nnn"}. Ignored with a
warning for files that already have a header.}

\item{purpose, input_files, output_files}{purpose, input files, and output
files of R program(s) given as \code{character} vectors. The purpose will
be wrapped to fit in the header. When \code{NULL} (the default), a blank
section is included. When \code{FALSE}, no section is included. Ignored
with a warning for files that already have a header.}

\item{backup}{\code{logical} indicating whether to create backup files}

\item{open}{\code{logical} indicating whether to open files in RStudio}
}
\value{
invisibly returns \code{NULL}
}
\description{
Create, update, and open R programs with QMS approved headers.
}
\details{
Cognigen's R program header documents key elements of programs including:
\itemize{
\item Name: The full path to the program on Cognigen's file system.
\item Timestamp: The timestamp and programmer of key updates to the program.
\item Copyright: Cognigen's standard Copyright language.
\item Purpose: What the program is designed to do.
\item Input Files: Files read by the program.
\item Output Files: Files written by the program.
}

\code{Redit} serves a few distinct purposes:
\itemize{
\item Create new R and Rmd files with standard headers
\item Add headers to existing R and Rmd files without a header
\item Add new timestamps to headers in existing R and Rmd files
}

\code{Redit} does not need to be used every time a program is updated and \code{Redit}
should not be used for files that are simply being opened.

If a standard header cannot be identified in an existing file, a new header
will be added if possible. In some cases, an error will occur with a
suggestion to update the program: For R files, update the first line so it is
an R command; For Rmd files, delete or rename any chunks named "header".
}
\examples{
\dontrun{
# Create two new R programs
Redit("file1.R", "file2.R")

# Update these programs and create a backup
Redit("file1.R", "file2.R", backup = TRUE)

# Update the timestamp of the open R program if it has a valid header.
# If there is no valid header, either a header will be added or the error
# will try to help resolve the issue
Redit()

# R Markdown works too
Redit("new-markdown.Rmd")

# Header items can be set programmatically, which is particularly useful when
# several programs are required
Redit(
  "new-file.R",
  purpose = "experiment with headers",
  input_files = c("../data/file1.csv", "..data/file1.xpt"),
  output_files = FALSE
)
}

}
\seealso{
\code{\link{Rcopy}} for copying R programs, \code{\link{get_header_content}}
for getting content from headers
}
