% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Render an R Markdown document and open the output}
\usage{
render(path = NULL, open = rstudioapi::isAvailable())
}
\arguments{
\item{path}{file path of R Markdown document. Defaults to the path of the
source editor context.}

\item{open}{\code{logical} indicating whether to open the output file}
}
\value{
\code{logical} indicating whether the document was successfully
  rendered.
}
\description{
An R program is generated with the name of \code{path} sans
extension, appended with "-render.R". This contains a header and a call to
\code{rmarkdown::}\code{\link[rmarkdown]{render}}. This program is then
executed from its directory and the output file is optionally opened.

The generated R program is not modified if it already exists, but the program
is examined to confirm it calls \code{rmarkdown::render}.
}
\examples{
\dontrun{
render("markdown_doc.Rmd")
}
}
