% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_roxygen.R
\name{render_roxygen}
\alias{render_roxygen}
\title{Render R documentation for roxygenized content}
\usage{
render_roxygen(
  ...,
  outdir = tempdir(),
  selected = FALSE,
  open = rstudioapi::isAvailable()
)
}
\arguments{
\item{...}{file paths to R files documented with \pkg{roxygen2}. Defaults to
the path of the source editor context.}

\item{outdir}{existing directory to output files to}

\item{selected}{\code{logical} indicating whether to include selected content
from the RStudio Source Editor. When \code{TRUE}, the path to the source
editor context is not automatically rendered in full.}

\item{open}{\code{logical} indicating whether to open the output files}
}
\value{
a named \code{logical} vector. The names represent the output files
  while the value indicates whether the output file was successfully
  rendered.
}
\description{
The main output file for each element of \code{...} contains links to each
help topic found in the file. This function is designed to render documentation
for functions documented with \pkg{roxygen2}.

Note that the code in provided files will be executed. This function is not
intended to be used on entire programs used for analysis or data assembly. To
only render documentation for a highlighted selection, see the
\code{selected} argument.
}
