% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanlogs.R
\name{dir_scanlogs}
\alias{dir_scanlogs}
\alias{scanlogs}
\title{Scan log files of all R and/or SAS programs within a directory}
\usage{
dir_scanlogs(path = getwd(), pattern = NULL, ext = c("Rout", "log"))

scanlogs(...)
}
\arguments{
\item{path}{a single directory path. Defaults to the working directory.}

\item{pattern}{optional \code{character} string containing a regular
expression. Only file names which match the regular expression will be
scanned}

\item{ext}{a \code{character} vector of file extensions to scan the logs of}

\item{...}{\code{character} vectors containing file names or paths with
extensions .R, .r, .Rout, .sas, .log, or .lst. Defaults to the path of the
source editor context.}
}
\value{
\code{dir_scanlogs()} invisibly returns a named \code{list} of
  \code{character} vectors

\code{scanlogs()} invisibly returns a \code{character} vector of
  scanned results or a \code{list} of \code{character} vectors when multiple
  paths are provided.
}
\description{
Scan the log files of multiple R and/or SAS programs The scan searches
\*.Rout files for warnings and errors; \*.log files for warnings, errors,
notes, and other messages using the system command \code{scanlogs}.
}
\examples{
\dontrun{
# Scan all Rout and SAS log files in the working directory
dir_scanlogs()

# Scan all Rout files with "final" in the file name
dir_scanlogs(path = "/stage_directory/asmbdat/", pattern = "final", ext = "Rout")
}

\dontrun{
scanlogs()
scanlogs("/stage_directory/asmbdat/mk-something.Rout")
scanlogs("/stage_directory/sas/check-something.log")
scanlogs("/stage_directory/asmbdat/mk-something.Rout", "/stage_directory/sas/check-something.log")
}

}
