% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanlogs.R
\name{scanlogs_single}
\alias{scanlogs_single}
\title{Scan log file of R, Rmd, or SAS programs}
\usage{
scanlogs_single(path = NULL)
}
\arguments{
\item{path}{a file name or path with extension .R, .r, .Rmd, .Rout, .r.Rout,
.sas, .log, or .lst. Defaults to the path of the source editor context.}
}
\value{
invisibly returns a \code{character} vector of scanned results. For
  R, the findings are prepended by the basename of the log path and line
  number. For SAS, the findings are prepended by the log path.
}
\description{
Only a single path should be passed directly to \code{scanlogs_single()}.
}
\keyword{internal}
