% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_code.R
\name{shared_code}
\alias{shared_code}
\alias{path_shared_code}
\alias{browse_shared_code}
\alias{list_files_shared_code}
\alias{Rcopy_shared_code}
\title{Interact with a shared code directory on the file system.}
\usage{
path_shared_code()

browse_shared_code()

list_files_shared_code()

Rcopy_shared_code(
  from,
  to = NULL,
  version = NULL,
  copyright_holder = NULL,
  open = rstudioapi::isAvailable()
)
}
\arguments{
\item{from}{relative path of an R program in the shared code directory.}

\item{to}{path or directory of new file. If \code{to} is a directory,
\code{from} is copied to that directory. Defaults to the basename of the
\code{from} file.}

\item{version}{either \code{NULL} for the current R version, or a
\code{character} in the form \code{"N.n.n"} or \code{"Nnn"}.}

\item{copyright_holder}{either \code{NULL} for the default Simulations Plus
copyright statement, a single \code{character} defining the copyright
holders and accompanying text to follow copyright mark and year, a
\code{character} vector for multiple separate copyright statements, or
\code{FALSE} for no copyright.}

\item{open}{\code{logical} indicating whether to open files in RStudio.}
}
\description{
Interact with a shared code directory on the file system.
}
\examples{
\dontrun{
# set the option that points to the shared code directory
options(utilscognigen.path_shared_code = "/path/to/shared-code")

# access the shared code directory path
path_shared_code()

# list files to see what's available
list_files_shared_code()

# or browse
browse_shared_code()

# copy a program with a new name
Rcopy_shared_code("templates/qc/dm-striking-qc.R", "qc-my-dataset.R")

# keep the original name
Rcopy_shared_code("templates/qc/dm-striking-qc.R")

# shorthand matching of file names is also supported
Rcopy_shared_code("dm-striking-qc")

# write to a different directory
Rcopy_shared_code("dm-striking-qc", "../asmbdat")

}
}
