% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{snippets}
\alias{snippets}
\alias{set_rstudio_snippets}
\title{Set and open RStudio snippets}
\usage{
set_rstudio_snippets(
  ...,
  include_rstudio_defaults = TRUE,
  include_shared_code = TRUE,
  include_ask_first = FALSE
)
}
\arguments{
\item{...}{named \code{character} vectors of file paths to snippet files. The
names may be any of the values of \code{type} (up to one argument per
\code{type}).}

\item{include_rstudio_defaults}{\code{logical} indicating whether to include
RStudio version 2022.2.3.492.3 default snippets.}

\item{include_shared_code}{\code{logical} indicating whether to include
snippets found in a "snippets" directory under
\code{\link{path_shared_code}()}.}

\item{include_ask_first}{\code{logical} indicating whether to ask which 
files to include from RStudio default snippets and shared code snippets.}
}
\value{
path to the users snippet file(s), invisibly.
}
\description{
\code{set_rstudio_snippets()} combines multiple snippet files, then opens
any modified snippet configs. The default behavior is to copy all of the 
default RStudio snippet types as well as all recognized snippet files found in
the \code{\link{path_shared_code}()} directory.

Files created by \code{\link[usethis]{edit_rstudio_snippets}()} and
\code{set_rstudio_snippets()} will *mask*, not supplement, the built-in
default snippets. If you like the built-in snippets, make sure they are
included with your custom snippets.
}
