% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_snapshots.R
\name{time_as_numeric}
\alias{time_as_numeric}
\title{Convert character representations of time to a numeric in the interval (0, 24].}
\usage{
time_as_numeric(time)
}
\arguments{
\item{time}{\code{character} vector of times assumed to be in 24-hour format;
will be adjusted if "AM" or "PM" is included.}
}
\value{
\code{numeric} vector of times in the interval (0, 24]
}
\description{
Convert character representations of time to a numeric in the interval (0, 24].
}
\keyword{internal}
