% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_path.R
\name{toggle_path_selection}
\alias{toggle_path_selection}
\title{Toggle paths between Unix or Windows based on the first entry}
\usage{
toggle_path_selection()
}
\description{
This is intended to be used as an RStudio Addin.
}
\keyword{internal}
