% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_aliases.R
\name{wd_aliases}
\alias{wd_aliases}
\alias{cd}
\alias{pwd}
\title{Change or print working directory}
\usage{
cd(dir)

pwd()
}
\arguments{
\item{dir}{A character string: \link[base]{tilde expansion} will be done.}
}
\description{
These are straight wrappers around \code{\link[base]{setwd}} and
\code{\link[base]{getwd}}.
}
