% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_source_file.R
\name{get_source_file}
\alias{get_source_file}
\title{Get the path of the RStudio source editor}
\usage{
get_source_file()
}
\value{
returns the path of the source editor; fails if there is no file
  active in the source editor.
}
\description{
Calls \code{\link[rstudioapi]{getSourceEditorContext}} to obtain the source
context.
}
