% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile.R
\name{logfile}
\alias{logfile}
\title{Identify and open R, Rmd, SAS, tex, Rnw log files}
\usage{
logfile(..., open = TRUE)
}
\arguments{
\item{...}{file paths to retrieve the log files of. Defaults to the path of
the source editor context.}

\item{open}{\code{logical} indicating whether to open files in RStudio.}
}
\value{
file paths to log files.
}
\description{
Identify and open R, Rmd, SAS, tex, Rnw log files
}
