% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Redit.R
\name{make_header}
\alias{make_header}
\title{Create a file with a header or append a modification history entry to an
existing file with a header}
\usage{
make_header(
  path = NULL,
  version = NULL,
  purpose = NULL,
  input_files = NULL,
  output_files = NULL,
  backup = FALSE
)
}
\arguments{
\item{version}{either \code{NULL} for the current R version, or a
\code{character} in the form \code{"N.n.n"} or \code{"Nnn"}. Ignored with a
warning for files that already have a header.}

\item{purpose}{purpose, input files, and output
files of R program(s) given as \code{character} vectors. The purpose will
be wrapped to fit in the header. When \code{NULL} (the default), a blank
section is included. When \code{FALSE}, no section is included. Ignored
with a warning for files that already have a header.}

\item{input_files}{purpose, input files, and output
files of R program(s) given as \code{character} vectors. The purpose will
be wrapped to fit in the header. When \code{NULL} (the default), a blank
section is included. When \code{FALSE}, no section is included. Ignored
with a warning for files that already have a header.}

\item{output_files}{purpose, input files, and output
files of R program(s) given as \code{character} vectors. The purpose will
be wrapped to fit in the header. When \code{NULL} (the default), a blank
section is included. When \code{FALSE}, no section is included. Ignored
with a warning for files that already have a header.}

\item{backup}{\code{logical} indicating whether to create backup files}
}
\value{
invisibly returns \code{NULL}
}
\description{
Create a file with a header or append a modification history entry to an
existing file with a header
}
\keyword{internal}
