% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Redit.R
\name{make_purpose}
\alias{make_purpose}
\title{Make the header lines containing the purpose}
\usage{
make_purpose(purpose = NULL, split_length = .split_length)
}
\arguments{
\item{split_length}{\code{integer} representing how many characters to split
at.}
}
\value{
\code{character} or \code{NULL}
}
\description{
Make the header lines containing the purpose
}
\keyword{internal}
