% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_aliases.R
\name{wd_aliases}
\alias{wd_aliases}
\alias{cd}
\alias{pwd}
\title{Change or print working directory}
\usage{
cd(dir, focus = getOption("utilscognigen.cd_focus", TRUE))

pwd()
}
\arguments{
\item{dir}{A character string: \link[base]{tilde expansion} will be done.}

\item{focus}{\code{logical} indicating whether to show the working directory
in the RStudio Files pane.}
}
\description{
These are wrappers around \code{\link[base]{setwd}} and
\code{\link[base]{getwd}}.

\code{cd()}, when called without any arguments, sets the working directory
to the directory of the source file.
}
